# src/main/resources/openapi/hearings.yaml
openapi: 3.0.1

info:

  version: ${version}
  title: Modernised DARTS
  description: |-
    Modernised DARTS (Digital Audio Recording and Transcription Service).

servers:
  - url: http://localhost:4550/


paths:
  /hearings/{hearing_id}:
    get:
      tags:
        - Hearings
      summary: Allows the retrieval of a hearing by ID.
      operationId: getHearing
      parameters:
        - in: path
          name: hearing_id
          schema:
            type: integer
          description: "Internal hea_id of the hearing."
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/get_hearing_response'
        '401':
          description: Unauthorised Error
        '403':
          description: Forbidden Error
          content:
            application/json+problem:
              schema:
                allOf:
                  - $ref: './problem.yaml'
                  - type: object
                    required:
                      - type
                    properties:
                      type:
                        $ref: '#/components/schemas/HearingAuthorisation403ErrorCode'
              example:
                type: "AUTHORISATION_100"
                title: "User is not authorised for the associated courthouse"
                status: 403
        '500':
          description: Internal Server Error
          content:
            application/json+problem:
              schema:
                $ref: './problem.yaml'
  /hearings/{hearing_id}/annotations:
    get:
      tags:
        - Hearings
      summary: Allows the retrieval of annotations for a hearing.
      operationId: getHearingAnnotations
      parameters:
        - in: path
          name: hearing_id
          schema:
            type: integer
          description: "Internal hea_id of the hearing."
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/annotation'
        '401':
          description: Unauthorised Error
        '403':
          description: Forbidden Error
          content:
            application/json+problem:
              schema:
                allOf:
                  - $ref: './problem.yaml'
                  - type: object
                    required:
                      - type
                    properties:
                      type:
                        $ref: '#/components/schemas/HearingAuthorisation403ErrorCode'
              example:
                type: "AUTHORISATION_100"
                title: "User is not authorised for the associated courthouse"
                status: 403
        '500':
          description: Internal Server Error
          content:
            application/json+problem:
              schema:
                $ref: './problem.yaml'
  /hearings/{hearing_id}/events:
    get:
      tags:
        - Hearings
      summary: Allows the retrieval of events for a given hearing_id
      operationId: getEvents
      description: |-
        Retrieves all events for a given hearing_id.
      parameters:
        - in: path
          name: hearing_id
          schema:
            type: integer
          description: "hearing_id is the internal hea_id of the hearing."
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/events_response'
        '500':
          description: Internal Server Error
          content:
            application/json+problem:
              schema:
                $ref: './problem.yaml'
  /hearings/{hearing_id}/transcripts:
    get:
      tags:
        - Hearings
      summary: Allows the retrieval of transcripts for a given hearing_id
      description: |-
        Retrieves all transcripts for a hearing_id.
      parameters:
        - in: path
          name: hearing_id
          schema:
            type: integer
          description: "hearing_id is the internal hea_id of the case."
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'cases.yaml#/components/schemas/transcripts'
        '401':
          description: Unauthorised Error
        '403':
          description: Forbidden Error
          content:
            application/json+problem:
              schema:
                allOf:
                  - $ref: './problem.yaml'
                  - type: object
                    required:
                      - type
                    properties:
                      type:
                        $ref: '#/components/schemas/HearingsAuthorisation403ErrorCode'
              example:
                type: "AUTHORISATION_100"
                title: "User is not authorised for the associated courthouse"
                status: 403
        '404':
          description: Not Found Error
          content:
            application/json+problem:
              schema:
                $ref: './problem.yaml'
              example:
                type: "CASE_104"
                title: "The requested hearing cannot be found"
                status: 404
        '500':
          description: Internal Server Error
          content:
            application/json+problem:
              schema:
                $ref: './problem.yaml'


components:
  schemas:
    get_hearing_response:
      type: object
      properties:
        hearing_id:
          type: integer
          example: 1
        courthouse:
          type: string
          example: SWANSEA
        courtroom:
          type: string
          example: 1
        hearing_date:
          type: string
          format: date
          example: 2023-06-23
        case_id:
          type: integer
          example: 1
        case_number:
          type: string
          example: CASE1001
        judges:
          type: array
          items:
            $ref: '#/components/schemas/judge'
        transcription_count:
          type: integer
          example: 1
        case_reporting_restrictions:
          $ref: './common.yaml#/components/schemas/reporting_restrictions'

    judge:
      type: string
      example: Mr Judge

    events_response:
      type: array
      default: [ ]
      items:
        $ref: '#/components/schemas/event_response'
    event_response:
      type: object
      properties:
        id:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
          example: 2023-07-31T14:32:24.620Z
        name:
          type: string
          example: Case called on
        text:
          type: string
          example: Record:New Case


    HearingAuthorisation403ErrorCode:
      type: string
      enum:
        - "AUTHORISATION_100"
      x-enum-varnames: [ USER_NOT_AUTHORISED_FOR_COURTHOUSE ]
    transcripts:
      type: array
      items:
        $ref: '#/components/schemas/transcript'
    transcript:
      type: object
      properties:
        transcription_id:
          type: integer
          example: 1
        hearing_id:
          type: integer
          example: 2
        hearing_date:
          type: string
          format: date
          example: 2022-05-20
        type:
          type: string
          example: "Sentencing remarks"
        requested_on:
          type: string
          format: date-time
          example: 2023-06-20T10:00:00Z
        requested_by_name:
          type: string
          example: "Joe Bloggs"
        status:
          type: string
          example: "Available"
    annotation:
      type: object
      properties:
        annotation_id:
          type: integer
          example: 1
        hearing_id:
          type: integer
          example: 2
        hearing_date:
          type: string
          format: date
          example: 2022-05-20
        annotation_ts:
          type: string
          format: date-time
          example: 2023-06-20T10:00:00Z
        annotation_text:
          type: string
          example: "A summary notes of this annotation..."
        annotation_documents:
          type: array
          items:
            $ref: '#/components/schemas/annotation_document'
    annotation_document:
      type: object
      properties:
        annotation_document_id:
          type: integer
          example: 1
        file_name:
          type: string
          example: "Annotation.doc"
        file_type:
          type: string
          example: "DOC"
        uploaded_by:
          type: string
          example: "Mr User McUserFace"
        uploaded_ts:
          type: string
          format: date-time
          example: 2023-06-20T10:00:00Z
    HearingsAuthorisation403ErrorCode:
      type: string
      enum:
        - "AUTHORISATION_100"
      x-enum-varnames: [ USER_NOT_AUTHORISED_FOR_COURTHOUSE ]